% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{organization_query}
\alias{organization_query}
\title{返回列表}
\usage{
organization_query(dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3")
}
\arguments{
\item{dms_token}{口令}
}
\value{
返回值
}
\description{
返回列表
}
\examples{
organization_query()
}
